Auth

+++ server/app/routes/orders/index.js
@@ -2,6 +2,8 @@ var router = require('express').Router();
 var bodyParser = require('body-parser');
 var mongoose = require('mongoose');
 var Order = mongoose.model('Order');
+var User = mongoose.model('User');
+var _ = require('lodash');
 
 module.exports = router;
 
@@ -11,29 +13,63 @@ function hasAdminPower(req, res, next){
 }
 
 // get all Orders
-router.get('/', hasAdminPower, function (req, res, next){
-	Order.find({}).exec()
-	.then(
-		function (orders){
-			res.json(orders);
-		}, 
-		function (err){
-			next(err);
-		}
-	);
+router.get('/', function (req, res, next){
+	if (req.user.admin) {
+		Order.find({}).exec()
+		.then(
+			function (orders){
+				res.json(orders);
+			}, 
+			function (err){
+				next(err);
+			}
+		);
+	} else {
+		console.log("only this user's orders");
+		User.findById(req.user._id)
+		.populate('Orders')
+		.exec()
+		.then(
+			function (user){
+				res.json(user.orders);
+			},
+			function (err){
+				next(err);
+			}
+		);
+	}
 });
 
 //get order with orderid 
 router.get('/:orderid', function (req, res, next){
-	Order.findById(req.params.orderid).exec()
-	.then(
-		function (order){
-			res.json(order);
-		},
-		function (err){
-			next(err);
-		}
-	);
+	if (req.user.admin) {
+		Order.findById(req.params.orderid).exec()
+		.then(
+			function (order){
+				res.json(order);
+			},
+			function (err){
+				next(err);
+			}
+		);
+	} else {
+		User.findById(req.user._id)
+		.populate('orders')
+		.exec()
+		.then(
+			function (user){
+				if (_.includes(user.orders, req.params.orderid)) {
+					var theOrder = _.find(user.orders, {_id: req.params.orderid})
+					res.json(theOrder);
+				} else {
+					res.sendStatus(403).end();
+				}
+			},
+			function (err){
+				next(err);
+			}
+		);
+	}
 });
 
 // creates new order and returns new order
@@ -46,26 +82,51 @@ router.post('/', function (req, res, next){
 
 //edits this order
 router.put('/:orderid', function (req, res, next){
-	Order.findByIdAndUpdate(req.params.orderid, req.body).exec()
-	.then(
-		function (order){
-			res.status(200).send(order);
-		},
-		function (err){
+	if (req.user.admin){
+		Order.findById(req.params.orderid)
+		.exec()
+		.then(function (order) {
+			order.changeStatus(req.body.orderstatus);
+			order.save();
+			res.json(order);
+			console.log(order);
+		}, function(err){
 			next(err);
-		}
-	);
+		});
+
+	} else {
+		User.findById(req.user._id)
+		.populate('orders')
+		.exec()
+		.then(
+			function (user){
+				if (_.includes(user.orders, req.params.orderid)) {
+					var theOrder = _.find(user.orders, {_id: req.params.orderid});
+					theOrder.cancelOrder();
+					theOrder.save();
+					res.json(theOrder);
+				} else {
+					res.sendStatus(403).end();
+				}
+			},
+			function (err){
+				next(err);
+			}
+		);
+	}
 });
 
 // delete this order
 router.delete('/:orderid', function (req, res, next){
-	Order.findByIdAndRemove(req.params.orderid).exec()
-	.then(
-		function (){
-			res.status(204).send();
-		},
-		function (err){
-			next(err);
-		}
-	);
+	if (req.user.admin){
+		Order.findByIdAndRemove(req.params.orderid).exec()
+		.then(
+			function (){
+				res.status(204).send();
+			},
+			function (err){
+				next(err);
+			}
+		);
+	}
 });
